generic ordering for balassetset
